Reformatted the fact checker agent for more agency + Created docker script for PostgreSQL db
Fact checker agent now analyses the claim, then passes it to the agent to conduct its research with a variety of tools, including calling the RAG bot for clarifications. Then, it saves the claim to DB, and upon recommendation from the agent, can pass it to the material generation agent.

Also, the DB is initiated automatically with docker.